covid/requirements.txt
asgiref==3.3.1
astroid==2.4.2
boto3==1.16.25
botocore==1.19.25
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
decorator==4.4.2
Django==3.1.3
django-rest-knox==4.1.0
djangorestframework==3.12.2
dynaconf==3.1.2
hvac==0.10.5
idna==2.10
isort==5.6.4
Jinja2==2.11.2
jmespath==0.10.0
knox==0.1.11
lazy-object-proxy==1.4.3
loguru==0.5.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
psycopg2==2.8.6
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
s3transfer==0.3.3
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
validators==0.18.1
wrapt==1.12.1
